Move shell_paths to common
