+++ cobalt/app/cobalt.cc
@@ -30,7 +30,7 @@
 #include "cobalt/app/cobalt_main_delegate.h"
 #include "cobalt/app/cobalt_switch_defaults_starboard.h"
 #include "cobalt/shell/browser/shell.h"
-#include "cobalt/shell/browser/shell_paths.h"
+#include "cobalt/shell/common/shell_paths.h"
 #include "content/public/app/content_main.h"
 #include "content/public/app/content_main_runner.h"
 #include "starboard/event.h"
