+++ cobalt/shell/common/shell_paths.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "cobalt/shell/browser/shell_paths.h"
+#include "cobalt/shell/common/shell_paths.h"
 
 #include "base/base_paths.h"
 #include "base/environment.h"
